+++ src/gui/web-components/structure/AppWebComponent.ts
@@ -1,6 +1,7 @@
 import { HeaderWebComponent } from "./HeaderWebComponent.js";
 import { MainWebComponent } from "./MainWebComponent.js";
 import { NotificationWebComponent } from "../util/NotificationWebComponent.js";
+import { FooterWebComponent } from "./FooterWebComponent.js";
 
 export class AppWebComponent extends HTMLElement {
     private isInitialized = false;
@@ -9,12 +10,14 @@ export class AppWebComponent extends HTMLElement {
         NotificationWebComponent.registerElement();
         HeaderWebComponent.registerElement();
         MainWebComponent.registerElement();
+        FooterWebComponent.registerElement();
 
         this.classList.add("platform-" + window.process.platform);
         this.innerHTML = `\
 <startgeraet-notification></startgeraet-notification>
 <startgeraet-header></startgeraet-header>
 <startgeraet-main></startgeraet-main>
+<startgeraet-footer></startgeraet-footer>
         `;
     }