+++ src/mcp/registry/registerConfigurationHandlers.ts
@@ -39,7 +39,7 @@ export function registerConfigurationHandlers(server: McpServer, config: Config)
                     .string()
                     .optional()
                     .describe(
-                        `Username or UserID for SimBrief API. Only required if user asks to populate Aerofly FS 4 by importing a Simbrief flightplan.`,
+                        `Username or UserID for SimBrief API. Only required if user asks to populate Aerofly FS 4 by importing a SimBrief flightplan.`,
                     ),
             },
             annotations: {