+++ types/gui/web-components/structure/FooterWebComponent.d.ts
@@ -1,6 +1,5 @@
 export declare class FooterWebComponent extends HTMLElement {
     private isInitialized;
-    private elements;
     private initialize;
     connectedCallback(): Promise<void>;
     static registerElement(): void;